Type mismatch AssertInt(x, y, op, er) in api.c

+++ tests/unit.h
@@ -46,8 +46,8 @@
 } while(0)
 
 #define AssertInt(x, y, op, er) do {                                           \
-    int _x = x;                                                                \
-    int _y = y;                                                                \
+    int _x = (int)x;                                                                \
+    int _y = (int)y;                                                                \
                                                                                \
     Assert(_x op _y, ("%s " #op " %s", #x, #y), ("%d " #er " %d", _x, _y));    \
 } while(0)
